+++ my-app/src/Components/Admin/main.css
@@ -779,16 +779,6 @@ select.form-control:focus::-ms-value {
   margin-bottom: 1rem;
 }
 
-.app{
-  min-height: calc(100vh - 50px);
-  padding: 10px;
-  margin-top: 30px;
-  background-color: #f5f5f5;
-  -webkit-transition: margin-left 0.3s ease;
-  -o-transition: margin-left 0.3s ease;
-  transition: margin-left 0.3s ease;
-} 
-
 .btn {
   display: inline-block;
   font-weight: 500;
@@ -1353,7 +1343,6 @@ div.dataTables_wrapper div.dataTables_paginate ul.pagination {
 @media (min-width: 768px) {
   .app-content {
     margin-left: 250px;
-    margin-top: 10px;
   }
 }
 
@@ -1409,113 +1398,6 @@ div.dataTables_wrapper div.dataTables_paginate ul.pagination {
 .app-header__logo:focus, .app-header__logo:hover {
   text-decoration: none;
   color: #fff;
-}
-
-.app-sidebar {
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  padding-top: 20px;
-  width: 250px;
-  padding-left: 10px;
-  padding-right: 10px;
-  overflow: auto;
-  z-index: 10;
-  background: rgb(0, 28, 64);
-}
-
-.app-sidebar::-webkit-scrollbar {
-  width: 6px;
-}
-
-.app-sidebar::-webkit-scrollbar-thumb {
-  background: rgba(0, 0, 0, 0.2);
-}
-
-.app-sidebar__user {
-  display: block;
-  display: -ms-flexbox;
-  -ms-flex-align: center;
-  align-items: center;
-  padding: 10px;
-  color: white;
-  text-align: center;
-}
-
-.app-sidebar__user-avatar {
-  display: inline-block;
-  background: white;
-  width: 100px;
-  height: 100px;
-  border: 3px solid white;
-  margin-bottom: 10px;
-  border-radius: 50%;
-}
-
-.app-sidebar__user-name {
-  font-size: 17px;
-  line-height: 1.3;
-}
-
-.app-sidebar__user-name,
-.app-sidebar__user-designation {
-  white-space: nowrap;
-  overflow: hidden;
-  -o-text-overflow: ellipsis;
-     text-overflow: ellipsis;
-  margin-bottom: 0;
-}
-
-.app-menu {
-  margin-bottom: 0;
-  padding-bottom: 40px;
-}
-
-.app-menu__item {
-  position: relative;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-  -ms-flex-align: center;
-  align-items: center;
-  border-radius: .375rem;
-  margin-bottom: 10px;
-  padding: 15px 15px;
-  font-size: 1.08em;
-  border-left: 0px;
-  box-shadow: none;
-  -webkit-transition: border-left-color 0.3s ease, background-color 0.3s ease;
-  -o-transition: border-left-color 0.3s ease, background-color 0.3s ease;
-  transition: border-left-color 0.3s ease, background-color 0.3s ease;
-  color: white;
-  border: 1px solid transparent;
-}
-
-.app-menu__item.active, .app-menu__item:hover, .app-menu__item:focus {
-  background: #c6defd;
-  text-decoration: none;
-  color: rgb(22 22 72);
-  box-shadow: none;
-  border: 1px solid rgb(22 22 72);
-}
-
-.app-menu__icon {
-  font-size: 23px;
-  -webkit-box-flex: 0;
-  -ms-flex: 0 0 auto;
-  flex: 0 0 auto;
-  width: 40px;
-}
-
-.app-menu__label {
-  white-space: nowrap;
-  -webkit-box-flex: 1;
-  -ms-flex: 1 1 auto;
-  flex: 1 1 auto;
-  font-size: 15px;
-  font-weight: 500;
 }
 
 .app-title {
